+++ internal/storage/mysql/permissions.go
@@ -16,7 +16,7 @@ func NewPermissionsStorage(db *sql.DB) *PermissionsStorage {
 	}
 }
 
-func (p *PermissionsStorage) Save(ctx context.Context, userId int, value int32) error {
+func (p *PermissionsStorage) Save(ctx context.Context, userId int64, value int32) error {
 	const op = "PermissionsStorage.Save"
 	if _, err := p.db.ExecContext(ctx, "INSERT INTO permissions (user_id, permission) VALUES (?, ?)", userId, value); err != nil {
 		return fmt.Errorf("%s: %w", op, err)
@@ -24,7 +24,7 @@ func (p *PermissionsStorage) Save(ctx context.Context, userId int, value int32) 
 	return nil
 }
 
-func (p *PermissionsStorage) Get(ctx context.Context, userId int) (int32, error) {
+func (p *PermissionsStorage) Get(ctx context.Context, userId int64) (int32, error) {
 	const op = "PermissionsStorage.Get"
 	var perm int32
 	if err := p.db.QueryRowContext(ctx, "SELECT permission FROM permissions WHERE user_id=?", userId).Scan(&perm); err != nil {
@@ -33,13 +33,13 @@ func (p *PermissionsStorage) Get(ctx context.Context, userId int) (int32, error)
 	return perm, nil
 }
 
-func (p *PermissionsStorage) Update(ctx context.Context, userId int, value int32) error {
+func (p *PermissionsStorage) Update(ctx context.Context, userId int64, value int32) error {
 	const op = "PermissionsStorage.Update"
 	//todo
 	return nil
 }
 
-func (p *PermissionsStorage) Delete(ctx context.Context, userId int) error {
+func (p *PermissionsStorage) Delete(ctx context.Context, userId int64) error {
 	const op = "PermissionsStorage.Delete"
 	// todo
 	return nil
